update_9 octobre #5 add likes

+++ backend/controllers/post.js
@@ -96,7 +96,7 @@ exports.modifyPost = async (req, res, next) => {
             if ((post.userId === auth) && (user.isActive)) {
 
                 //========// TRAITEMENT DE LA REQUETE
-                let content = req.body
+                let content = {...req.body, updated : true}
 
                 //----Suppression Image
                 if (req.body.nofile === true) {
@@ -218,7 +218,7 @@ exports.deletePost = async (req, res, next) => {
     }
 };
 
-//================//LIKER//================//
+//================// LIKER //================//
 exports.likePost = async (req, res, next) => {
     const liker = req.auth.userId
     const postToLike = Number(req.params.id)
@@ -250,16 +250,57 @@ exports.likePost = async (req, res, next) => {
                     }
                 })
                 .then(async () => { await prisma.$disconnect() })
-                .then(() => res.status(201).json({ message : 'publication likee !'}))
+                
+
+                .then(async () => {
+                    await prisma.like.findFirst({
+                        where : {
+                            userId : liker,
+                            postId : postToLike
+                        }
+                    })
+                    .then((like) => res.status(201).json(like))
+                })
                 .catch(error =>  res.status(500).json(error))
             }
             // empecher autolike
             else {
-                return res.status(403).json(errMsg.likeErr)
+                return res.status(403).json(errMsg.autolikeErr)
             }
         }
-        //========//doublon : RETIRER
+        //========//doublon
         else {
+            return res.status(403).json(errMsg.likeErr)
+        }
+    // mauvais utilisateur
+    } else {
+        return res.status(403).json(errMsg.authErr);
+    }
+};
+
+//================// DISLIKE //================//
+exports.unlikePost = async (req, res, next) => {
+    const liker = req.auth.userId
+    const postToLike = Number(req.params.id)
+    
+    //---RECHERCHES
+    // utilisateur
+    const findUser = await utils.findUser({id : liker})
+    // publication
+    const findPost = await utils.findPost({id : postToLike})
+    // like doublon
+    const findLike = await prisma.like.findFirst({
+        where : {
+            userId : liker,
+            postId : postToLike
+        }
+    })
+
+    // l'utilisateur et le post doivent etre actif
+    if ((findUser.isActive) && (findPost.isActive)) {
+        // CONDITIONS
+        //========// RETIRER
+        if (Boolean(findLike)) {
             await prisma.like.delete({
                 where : { id : findLike.id }
             })
@@ -267,8 +308,12 @@ exports.likePost = async (req, res, next) => {
             .then(() => res.status(200).json({ message : 'like supprime !' }))
             .catch(error => res.status(500).json(error))
         }
+        else {
+            return res.status(400).json(errMsg.unlikeErr);
+        }
+
     // mauvais utilisateur
     } else {
         return res.status(403).json(errMsg.authErr);
     }
-};
+}
